src/requirements.txt
ansible-core==2.14.3 ; python_version >= "3.10" and python_version < "4.0"
ansible==7.3.0 ; python_version >= "3.10" and python_version < "4.0"
ansicon==1.89.0 ; python_version >= "3.10" and python_version < "4" and platform_system == "Windows"
arrow==1.2.3 ; python_version >= "3.10" and python_version < "4"
asgiref==3.6.0 ; python_version >= "3.10" and python_version < "4.0"
blessed==1.20.0 ; python_version >= "3.10" and python_version < "4"
certifi==2022.12.7 ; python_version >= "3.10" and python_version < "4"
cffi==1.15.1 ; python_version >= "3.10" and python_version < "4.0"
charset-normalizer==3.1.0 ; python_version >= "3.10" and python_version < "4"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "4.0" and platform_system == "Windows"
crispy-bulma==0.8.3 ; python_version >= "3.10" and python_version < "4.0"
cryptography==39.0.2 ; python_version >= "3.10" and python_version < "4.0"
debugpy==1.6.6 ; python_version >= "3.10" and python_version < "4.0"
django-cleanup==7.0.0 ; python_version >= "3.10" and python_version < "4.0"
django-crispy-forms==2.0 ; python_version >= "3.10" and python_version < "4.0"
django-extensions==3.2.1 ; python_version >= "3.10" and python_version < "4.0"
django-picklefield==3.1 ; python_version >= "3.10" and python_version < "4"
django-pictures==1.0.1 ; python_version >= "3.10" and python_version < "4.0"
django-prometheus==2.2.0 ; python_version >= "3.10" and python_version < "4.0"
django-q==1.3.9 ; python_version >= "3.10" and python_version < "4"
django==4.1.7 ; python_version >= "3.10" and python_version < "4.0"
gunicorn==20.1.0 ; python_version >= "3.10" and python_version < "4.0"
idna==3.4 ; python_version >= "3.10" and python_version < "4"
jinja2==3.1.2 ; python_version >= "3.10" and python_version < "4.0"
jinxed==1.2.0 ; python_version >= "3.10" and python_version < "4" and platform_system == "Windows"
markupsafe==2.1.2 ; python_version >= "3.10" and python_version < "4.0"
mysqlclient==2.1.1 ; python_version >= "3.10" and python_version < "4.0"
numpy==1.24.2 ; python_version < "4.0" and python_version >= "3.10"
packaging==23.0 ; python_version >= "3.10" and python_version < "4.0"
pandas==1.5.3 ; python_version >= "3.10" and python_version < "4.0"
pillow==9.4.0 ; python_version >= "3.10" and python_version < "4.0"
prometheus-client==0.16.0 ; python_version >= "3.10" and python_version < "4.0"
pycparser==2.21 ; python_version >= "3.10" and python_version < "4.0"
python-dateutil==2.8.2 ; python_version >= "3.10" and python_version < "4.0"
python-dotenv==1.0.0 ; python_version >= "3.10" and python_version < "4.0"
pytube==12.1.2 ; python_version >= "3.10" and python_version < "4.0"
pytz==2022.7.1 ; python_version >= "3.10" and python_version < "4.0"
pyyaml==6.0 ; python_version >= "3.10" and python_version < "4.0"
redis==3.5.3 ; python_version >= "3.10" and python_version < "4"
requests==2.28.2 ; python_version >= "3.10" and python_version < "4"
resolvelib==0.8.1 ; python_version >= "3.10" and python_version < "4.0"
setuptools==67.6.0 ; python_version >= "3.10" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.10" and python_version < "4"
sqlparse==0.4.3 ; python_version >= "3.10" and python_version < "4.0"
tqdm==4.65.0 ; python_version >= "3.10" and python_version < "4.0"
tzdata==2022.7 ; python_version >= "3.10" and python_version < "4.0" and sys_platform == "win32"
urllib3==1.26.15 ; python_version >= "3.10" and python_version < "4"
wcwidth==0.2.6 ; python_version >= "3.10" and python_version < "4"
